+++ src/onegov.election_day/onegov/election_day/cli.py
@@ -5,6 +5,7 @@ import os
 
 from onegov.core.cli import command_group, pass_group_context
 from onegov.election_day.models import ArchivedResult
+from onegov.election_day.utils import add_local_results
 from onegov.election_day.sms_processor import SmsQueueProcessor
 
 
@@ -41,33 +42,34 @@ def fetch(group_context):
     """
 
     def fetch_results(request, app):
+        if not app.principal:
+            return
+
         local_session = app.session()
         assert local_session.info['schema'] == app.schema
 
-        available = app.session_manager.list_schemas()
+        for key in app.principal.fetch:
+            schema = '{}-{}'.format(app.namespace, key)
+            assert schema in app.session_manager.list_schemas()
+            app.session_manager.set_current_schema(schema)
+            remote_session = app.session_manager.session()
+            assert remote_session.info['schema'] == schema
 
-        if app.principal:
-            for key in app.principal.fetch:
-                schema = '{}-{}'.format(app.namespace, key)
-                assert schema in available
-                app.session_manager.set_current_schema(schema)
-                remote_session = app.session_manager.session()
-                assert remote_session.info['schema'] == schema
+            items = local_session.query(ArchivedResult)
+            items = items.filter_by(schema=schema)
+            for item in items:
+                local_session.delete(item)
 
-                items = local_session.query(ArchivedResult)
-                items = items.filter_by(schema=schema)
+            for domain in app.principal.fetch[key]:
+                items = remote_session.query(ArchivedResult)
+                items = items.filter_by(schema=schema, domain=domain)
                 for item in items:
-                    local_session.delete(item)
-
-                for domain in app.principal.fetch[key]:
-                    items = remote_session.query(ArchivedResult)
-                    items = items.filter_by(schema=schema, domain=domain)
-                    for item in items:
-                        new_item = ArchivedResult()
-                        new_item.copy_from(item)
-                        local_session.add(new_item)
-
-        click.echo("Results fetched successfully")
+                    new_item = ArchivedResult()
+                    new_item.copy_from(item)
+                    add_local_results(
+                        item, new_item, app.principal, remote_session
+                    )
+                    local_session.add(new_item)
 
     return fetch_results